Move MatchingCuts to dedicated file and change include paths accordingly.
